+++ src/components/Header.vue
@@ -23,6 +23,7 @@ export default {
     window.gapi.signin2.render('google-signin-hook', {
       onsuccess: this.onSignIn
     })
+    console.log(this.account)
   },
   computed: {
     ...mapGetters(['account', 'profile'])
