--- src/store/getters.js
@@ -1,11 +1,0 @@
-export default {
-  account: state => state.account,
-  address: state => state.address,
-  metamask: state => state.metamask,
-  connected: state => state.connected,
-  Contract: state => state.Contract,
-  // w3Data: state => state.w3Data,
-  abi: state => state.abi,
-  items: state => state.items,
-  profile: state => state.profile
-}
