+++ src/store/index.js
@@ -1,16 +1,14 @@
 import Vue from 'vue'
 import Vuex from 'vuex'
-import actions from './actions'
-import getters from './getters'
-import mutations from './mutations'
-import state from './state'
+import web3init from './web3init'
+import oauth from './oauth'
 
 Vue.use(Vuex)
 
 export default new Vuex.Store({
-  state,
-  getters,
-  actions,
-  mutations,
-  strict: false
+  strict: false,
+  modules: {
+    web3: web3init,
+    oauth: oauth
+  }
 })
