--- src/store/state.js
@@ -1,23 +1,0 @@
-import abi from '../../build/contracts/Murdoch.json'
-
-export default {
-  retried: false,
-  metamask: false,
-  account: null,
-  connected: false,
-  Contract: null,
-  contractAddress: null,
-  rpcUrl: null,
-  abi,
-  items: [
-    {
-      ref: '0x06012c8cf97bead5deae237070f9587f8e7a266d',
-      id: '1066656'
-    },
-    {
-      ref: '0x06012c8cf97bead5deae237070f9587f8e7a266d',
-      id: '1066657'
-    }
-  ],
-  profile: {}
-}
